@@ -39,14 +39,13 @@
   "main": "./index.js",
   "dependencies": {
     "debug": "^2.2.0",
-    "immutability-helper": "^2.0.0",
     "immutable": "^3.8.1",
     "normalize.css": "^5.0.0",
-    "react": "^15.3.1",
-    "react-addons-css-transition-group": "^15.3.1",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
     "react-dnd": "^2.1.4",
     "react-dnd-html5-backend": "^2.1.2",
-    "react-dom": "^15.3.1",
+    "react-dom": "^15.6.1",
     "react-mdl": "^1.9.0",
     "react-modal": "^1.6.4",
     "react-redux": "^4.4.5",
@@ -77,7 +76,7 @@
     "jest": "^20.0.4",
     "node-sass": "^4.5.3",
     "prettier": "^1.6.0",
-    "react-test-renderer": "^15.4.2",
+    "react-test-renderer": "^15.6.1",
     "redux-devtools": "^3.3.1",
     "sass-loader": "^6.0.6",
     "style-loader": "^0.18.2",
