@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 
 import {
@@ -13,7 +14,7 @@ import { HeaderTitle } from '../common';
 
 class StrategiesListComponent extends Component {
     static contextTypes = {
-        router: React.PropTypes.object,
+        router: PropTypes.object,
     };
 
     componentDidMount() {
