@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 import { Switch, Chip, ListItem } from 'react-mdl';
 import Progress from './progress';
