@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { Menu, MenuItem, IconButton } from 'react-mdl';
 
 class AddStrategy extends React.Component {
