@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import HistoryList from './history-list-container';
 
 class HistoryListToggle extends Component {
