@@ -1,4 +1,5 @@
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 import ViewFeatureToggle from '../../component/feature/view-container';
 
 export default class Features extends PureComponent {
