@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import HistoryListToggle from '../../component/history/history-list-toggle-container';
 
 const render = ({ params }) => (
