@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import ShowStrategy from '../../component/strategies/strategy-details-container';
 
 const render = ({ params }) => (
